# API-doc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WebShop24",
    "description": "This is the best WebShop you can find.\n",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "https://petstore.swagger.io/v2"
    }
  ],
...
  "paths": {
...
    "/seller": {
      "post": {
        "tags": [
          "seller"
        ],
        "summary": "Create a new seller (registration)",
        "operationId": "createSeller",
        "requestBody": {
          "$ref": "#/components/requestBodies/Seller"
        },
        "responses": {
          "400": {
            "description": "User Data uncomplete"
          },
          "403": {
            "description": "The given Zip-Code doesn't exist or is not in the given city.\nSeller with this Email already exists.\nInvalid Email.\nPassword is not secure."
          },
          "201": {
            "description": "Seller created",
            "content": {
              "application/json": {
                "example": {
                  "_id": "cWGtlAnhcsVViidg",
                  "password": "Test123#",
                  "brand": "testBrand",
                  "email": "test@test.de",
                  "iban": "DE1234567890",
                  "address": "Leiser Weg 12",
                  "zipCode": 70569,
                  "city": "Stuttgart",
                  "updated-at":"2024-01-07T17:15:24.623Z",
                  "created-at": "2024-01-07T17:15:24.623Z"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "seller"
        ],
        "summary": "Get all sellers, not used in the frontend application",
        "operationId": "getSellers",
        "responses": {
          "500": {
            "description": "Internal Server Error."
          },
          "200": {
            "description": "Sellers returned",
            "content": {
              "application/json": {
                "example": [
                  {
                    "_id": "cWGtlAnhcsVViidg",
                    "password": "Test123#",
                    "brand": "testBrand",
                    "email": "test@test.de",
                    "iban": "DE1234567890",
                    "address": "Leiser Weg 12",
                    "zipCode": 70569,
                    "city": "Stuttgart",
                    "updated-at":"2024-01-07T17:15:24.623Z",
                    "created-at": "2024-01-07T17:15:24.623Z"
                  },
                  {
                    "_id": "cWGtlAnhcsVViidk",
                    "password": "Test123#",
                    "brand": "testBrand2",
                    "email": "test@test.de",
                    "iban": "DE1234567890",
                    "address": "Leiser Weg 12",
                    "zipCode": 70569,
                    "city": "Stuttgart",
                    "updated-at":"2024-01-07T17:15:24.623Z",
                    "created-at": "2024-01-07T17:15:24.623Z"
                  }
                ]
              }
            }
          }
        }
      }
    },
    "/seller/{sellerId}": {
      "get": {
        "tags": [
          "seller"
        ],
        "summary": "Get Seller by SellerID",
        "operationId": "getSellerByID",
        "requestBody": {
          "$ref": "#/components/requestBodies/Buyer"
        },
        "responses": {
          "404": {
            "description": "Seller not found"
          },
          "500": {
            "description": "Internal Server Error."
          },
          "200": {
            "description": "Seller returned",
            "content": {
              "application/json": {
                "example": {
                  "_id": "cWGtlAnhcsVViidg",
                  "password": "Test123#",
                  "brand": "testBrand",
                  "email": "test@test.de",
                  "iban": "DE1234567890",
                  "address": "Leiser Weg 12",
                  "zipCode": 70569,
                  "city": "Stuttgart",
                  "updated-at":"2024-01-07T17:15:24.623Z",
                  "created-at": "2024-01-07T17:15:24.623Z"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "seller"
        ],
        "summary": "Update Seller",
        "operationId": "updateSellerByID",
        "requestBody": {
          "$ref": "#/components/requestBodies/SellerPatch"
        },
        "responses": {
          "400": {
            "description": "User Data incomplete"
          },
          "403": {
            "description": "The given Zip-Code doesn't exist or is not in the given city.\nInvalid Email.\nPassword is not secure."
          },
          "401": {
            "description": "Wrong Password"
          },
          "404": {
            "description": "Wrong Username"
          },
          "200": {
            "description": "Seller updated",
            "content": {
              "application/json": {
                "example": {
                  "_id": "cWGtlAnhcsVViidg",
                  "password": "Test123#",
                  "brand": "testBrand",
                  "email": "test@test.de",
                  "iban": "DE1234567890",
                  "address": "Leiser Weg 12",
                  "zipCode": 70569,
                  "city": "Stuttgart",
                  "updated-at":"2024-01-07T17:15:24.623Z",
                  "created-at": "2024-01-07T17:15:24.623Z"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "seller"
        ],
        "summary": "Delete Seller",
        "operationId": "deleteSeller",
        "requestBody": {
          "$ref": "#/components/requestBodies/Buyer"
        },
        "responses": {
          "401": {
            "description": "Wrong Password"
          },
          "404": {
            "description": "Wrong Username"
          },
          "500": {
            "description": "Internal Server Error."
          },
          "200": {
            "description": "Seller deleted"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Buyer": {
        "required": [
          "firstName",
          "lastName",
          "email",
          "password",
          "iban",
          "address",
          "zipCode",
          "city"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "format": "BSON",
            "readOnly": true
          },
          "password": {
            "type": "string",
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@#$%^&+=!]).{8,}$",
            "example": "Test123#"
          },
          "firstName": {
            "type": "string",
            "example": "Tarek"
          },
          "lastName": {
            "type": "string",
            "example": "Bürner"
          },
          "email": {
            "type": "string",
            "example": "test@test.de"
          },
          "address": {
            "type": "string",
            "example": "Leiser Weg 12"
          },
          "zipCode": {
            "type": "number",
            "pattern": "5 digits, existing zipCode",
            "example": 70569
          },
          "city": {
            "type": "string",
            "pattern": "must match the zipCode",
            "example": "Stuttgart"
          }
        }
      },
      "Seller": {
        "required": [
          "email",
          "password",
          "address",
          "iban",
          "brand",
          "zipCode",
          "city"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "format": "BSON",
            "readOnly": true
          },
          "password": {
            "type": "string",
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@#$%^&+=!]).{8,}$",
            "example": "Test123#"
          },
          "brand": {
            "type": "string",
            "example": "Something"
          },
          "email": {
            "type": "string",
            "example": "test@test.de"
          },
          "iban": {
            "type": "string",
            "example": "DE1234567890"
          },
          "address": {
            "type": "string",
            "example": "Leiser Weg 12"
          },
          "zipCode": {
            "type": "number",
            "pattern": "5 digit number",
            "example": 70569
          },
          "city": {
            "type": "string",
            "pattern": "must match the zipCode",
            "example": "Stuttgart"
          }
        }
      },
      "SellerPatch": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "format": "BSON",
            "readOnly": true
          },
          "password": {
            "type": "string",
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@#$%^&+=!]).{8,}$",
            "example": "Test123#"
          },
          "brand": {
            "type": "string",
            "example": "Something"
          },
          "email": {
            "type": "string",
            "example": "test@test.de"
          },
          "iban": {
            "type": "string",
            "example": "DE1234567890"
          },
          "address": {
            "type": "string",
            "example": "Leiser Weg 12"
          },
          "zipCode": {
            "type": "number",
            "pattern": "5 digit number",
            "example": 70569
          },
          "city": {
            "type": "string",
            "pattern": "must match the zipCode",
            "example": "Stuttgart"
          }
        }
      },
...
    },
    "requestBodies": {
      "Buyer": {
        "description": "Buyer, that should be registered",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Buyer"
            }
          }
        },
        "required": true
      },
...
      "Seller": {
        "description": "Seller, that should be created",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Seller"
            }
          }
        },
        "required": true
      },
      "SellerPatch": {
        "description": "Seller, that should be updated",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SellerPatch"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
...
  }
}